Fix potential AbstractMethodError on KotlinJvmBinaryClass.getContainingLibrary()
Compiler supposed to generate such method inside the
LocalFileKotlinClass class, but because of the KT-18489 method
getContainingLibrary is not generated. Because of that we potentially
could get AbstractMethodError. In order to fix that we are adding
explicit ovverride in java super class

(Cherry-picked from 3802ef2815d6a93acdb849ce519d638171fe5c41)